Add events presenter
